refactor: move filters para backend/diabetes
