+++ src/features/acf/backend/diabetes/filters/adapter.ts
@@ -3,10 +3,7 @@ import type {
     PatientAgeRange,
     PatientStatus,
 } from "@/features/acf/diabetes/common/model";
-import type {
-    FilterOptionsDb,
-    FiltersOptions,
-} from "../../../backend/diabetes/model";
+import type { FilterOptionsDb, FiltersOptions } from "../model";
 
 // TODO mudar esses tipos para não serem hard coded
 export type DiabetesFilterOptions =
