+++ src/features/acf/backend/diabetes/filters/index.ts
@@ -1,0 +1,1 @@
+export { filterOptionsCoeq } from "./controller";
