+++ src/features/acf/backend/diabetes/index.ts
@@ -1,1 +1,4 @@
+//TODO: importar direto de ./table
 export { page as getPage, rowCount as getRowCount } from "./table/controller";
+
+export { filterOptionsCoeq } from "./filters";
